+++ apps/events-backend/prisma/schema/base.prisma
@@ -1,0 +1,7 @@
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+}
