+++ apps/events-backend/prisma/schema/users.prisma
@@ -1,0 +1,90 @@
+model User {
+  id               String   @id
+  email            String   @unique
+  name             String
+  identityDocument String?
+  academicId       String?
+  role             UserRole @default(USER)
+
+  people People[]
+
+  createdAt   DateTime @default(now())
+  createdById String?
+  updatedAt   DateTime @updatedAt
+  updatedById String?
+
+  @@index([identityDocument])
+  @@index([academicId])
+  @@map("users")
+}
+
+enum UserRole {
+  USER
+  EVENT_MANAGER
+  CACIC
+  ADMIN
+}
+
+enum ContactType {
+  EMAIL
+  PHONE
+  WHATSAPP
+  OTHER
+}
+
+model AccountUserMerge {
+  oldUserId String @id
+  newUserId String
+
+  createdAt DateTime @default(now())
+
+  @@index([newUserId])
+  @@map("account_user_merges")
+}
+
+enum ExternalAccountMergeStatus {
+  APPLIED
+  FAILED
+  ROLLED_BACK
+}
+
+enum ExternalAccountMergeResult {
+  PEOPLE_MERGED
+  PERSON_REASSIGNED
+  ALREADY_APPLIED
+  NO_LOCAL_PERSON
+}
+
+model ExternalAccountMergeOperation {
+  id String @id @default(uuid(7))
+
+  eventId    String @unique
+  type       String
+  oldUserId  String
+  newUserId  String
+  occurredAt DateTime
+
+  status ExternalAccountMergeStatus @default(APPLIED)
+  result ExternalAccountMergeResult?
+
+  peopleMergeOperationId String?
+  peopleMergeOperation   PeopleMergeOperation? @relation(fields: [peopleMergeOperationId], references: [id])
+
+  requestPayload Json
+  errorMessage   String?
+  attemptCount   Int     @default(1)
+
+  rolledBackAt   DateTime?
+  rolledBackById String?
+
+  createdAt   DateTime @default(now())
+  createdById String?
+  updatedAt   DateTime @updatedAt
+  updatedById String?
+
+  @@index([oldUserId])
+  @@index([newUserId])
+  @@index([status])
+  @@index([peopleMergeOperationId])
+  @@map("external_account_merge_operations")
+}
