+++ apps/events-backend/src/app/events/attendances/event-attendances.queries.resolver.ts
@@ -1,0 +1,250 @@
+import { EventAttendance, EventAttendanceScannerFeedItem, MajorEventUserAttendance } from '@cacic-fct/shared-data-types';
+import { NotFoundException } from '@nestjs/common';
+import { Args, Int, Query, Resolver } from '@nestjs/graphql';
+import { Prisma } from '@prisma/client';
+import { RequireScopes } from '../../auth/decorators/require-scopes.decorator';
+import { PrismaService } from '../../prisma/prisma.service';
+import { AttendanceCategoryService } from '../attendance-category.service';
+import { EventAttendancesResolverBase, EVENT_RELATION_SELECT } from './event-attendances.shared';
+
+@Resolver(() => EventAttendance)
+export class EventAttendancesQueriesResolver extends EventAttendancesResolverBase {
+  constructor(prisma: PrismaService, attendanceCategories: AttendanceCategoryService) {
+    super(prisma, attendanceCategories);
+  }
+
+  @Query(() => [EventAttendance], { name: 'eventAttendances' })
+  @RequireScopes('event-attendance#read')
+  async eventAttendances(
+    @Args('personId', { type: () => String, nullable: true }) personId?: string,
+    @Args('eventId', { type: () => String, nullable: true }) eventId?: string,
+    @Args('skip', { type: () => Int, nullable: true }) skip?: number,
+    @Args('take', { type: () => Int, nullable: true }) take?: number,
+  ) {
+    const where: Prisma.EventAttendanceWhereInput = {};
+
+    if (personId) {
+      where.personId = personId;
+    }
+
+    if (eventId) {
+      where.eventId = eventId;
+    }
+
+    const attendances = await this.prisma.eventAttendance.findMany({
+      where,
+      select: {
+        personId: true,
+        eventId: true,
+        attendedAt: true,
+        createdAt: true,
+        createdById: true,
+        createdByMethod: true,
+        collectedLatitude: true,
+        collectedLongitude: true,
+        collectedAccuracyMeters: true,
+        category: true,
+        person: true,
+        event: {
+          select: EVENT_RELATION_SELECT,
+        },
+      },
+      orderBy: {
+        attendedAt: 'desc',
+      },
+      skip,
+      take,
+    });
+
+    const collectorIds = [
+      ...new Set(attendances.map((attendance) => attendance.createdById).filter((id): id is string => Boolean(id))),
+    ];
+    const collectors = collectorIds.length
+      ? await this.prisma.user.findMany({
+          where: {
+            id: {
+              in: collectorIds,
+            },
+          },
+          select: {
+            id: true,
+            name: true,
+          },
+        })
+      : [];
+    const collectorNameById = new Map(collectors.map((collector) => [collector.id, collector.name]));
+
+    return attendances.map((attendance) => ({
+      ...attendance,
+      collectedByFullName: attendance.createdById ? (collectorNameById.get(attendance.createdById) ?? undefined) : undefined,
+    }));
+  }
+
+  @Query(() => [EventAttendanceScannerFeedItem], { name: 'eventAttendanceScannerFeed' })
+  @RequireScopes('event-attendance#read')
+  eventAttendanceScannerFeed(@Args('eventId', { type: () => String }) eventId: string) {
+    return this.getScannerFeed(eventId);
+  }
+
+  @Query(() => [MajorEventUserAttendance], {
+    name: 'majorEventUserAttendances',
+  })
+  @RequireScopes('event-attendance#read')
+  async majorEventUserAttendances(
+    @Args('majorEventId', { type: () => String }) majorEventId: string,
+    @Args('personId', { type: () => String, nullable: true }) personId?: string,
+    @Args('skip', { type: () => Int, nullable: true }) skip?: number,
+    @Args('take', { type: () => Int, nullable: true }) take?: number,
+  ) {
+    const majorEvent = await this.prisma.majorEvent.findFirst({
+      where: {
+        id: majorEventId,
+        deletedAt: null,
+      },
+      select: {
+        id: true,
+      },
+    });
+
+    if (!majorEvent) {
+      throw new NotFoundException(`Major event ${majorEventId} was not found.`);
+    }
+
+    const events = await this.prisma.event.findMany({
+      where: {
+        majorEventId,
+        deletedAt: null,
+      },
+      select: {
+        id: true,
+        name: true,
+        startDate: true,
+      },
+      orderBy: {
+        startDate: 'asc',
+      },
+    });
+
+    if (events.length === 0) {
+      return [];
+    }
+
+    const eventIds = events.map((event) => event.id);
+    const subscriptions = await this.prisma.majorEventSubscription.findMany({
+      where: {
+        majorEventId,
+        deletedAt: null,
+        ...(personId ? { personId } : {}),
+      },
+      include: {
+        person: {
+          include: {
+            user: true,
+          },
+        },
+      },
+      orderBy: {
+        createdAt: 'desc',
+      },
+      skip,
+      take,
+    });
+
+    const attendances = await this.prisma.eventAttendance.findMany({
+      where: {
+        eventId: {
+          in: eventIds,
+        },
+        ...(personId ? { personId } : {}),
+      },
+      select: {
+        personId: true,
+        eventId: true,
+        attendedAt: true,
+        category: true,
+        person: {
+          include: {
+            user: true,
+          },
+        },
+      },
+    });
+
+    const personIds = Array.from(
+      new Set([
+        ...subscriptions.map((subscription) => subscription.personId),
+        ...attendances.map((attendance) => attendance.personId),
+      ]),
+    );
+    const attendanceByKey = new Map(
+      attendances.map((attendance) => [`${attendance.personId}:${attendance.eventId}`, attendance]),
+    );
+
+    const majorSubscriptionByPerson = new Map(
+      subscriptions.map((subscription) => [subscription.personId, subscription]),
+    );
+
+    return personIds.map((resolvedPersonId) => {
+      const subscription = majorSubscriptionByPerson.get(resolvedPersonId);
+      const person =
+        subscription?.person ?? attendances.find((attendance) => attendance.personId === resolvedPersonId)?.person;
+
+      return {
+        majorEventId,
+        subscriptionId: subscription?.id,
+        personId: resolvedPersonId,
+        person,
+        subscriptionStatus: subscription?.subscriptionStatus ?? 'UNKNOWN',
+        amountPaid: subscription?.amountPaid,
+        paymentDate: subscription?.paymentDate,
+        paymentTier: subscription?.paymentTier,
+        attendances: events.map((event) => {
+          const attendance = attendanceByKey.get(`${resolvedPersonId}:${event.id}`);
+          return {
+            eventId: event.id,
+            eventName: event.name,
+            eventStartDate: event.startDate,
+            attended: attendance != null,
+            attendedAt: attendance?.attendedAt,
+            category: attendance?.category ?? 'UNKNOWN',
+          };
+        }),
+      };
+    });
+  }
+
+  @Query(() => EventAttendance, { name: 'eventAttendance' })
+  @RequireScopes('event-attendance#read')
+  async eventAttendance(
+    @Args('personId', { type: () => String }) personId: string,
+    @Args('eventId', { type: () => String }) eventId: string,
+  ) {
+    const attendance = await this.prisma.eventAttendance.findUnique({
+      where: {
+        personId_eventId: {
+          personId,
+          eventId,
+        },
+      },
+      select: {
+        personId: true,
+        eventId: true,
+        attendedAt: true,
+        createdAt: true,
+        createdById: true,
+        createdByMethod: true,
+        category: true,
+        person: true,
+        event: {
+          select: EVENT_RELATION_SELECT,
+        },
+      },
+    });
+
+    if (!attendance) {
+      throw new NotFoundException(`Attendance ${personId}/${eventId} was not found.`);
+    }
+
+    return attendance;
+  }
+}
