+++ apps/events-backend/src/app/events/attendances/event-attendances.shared.ts
@@ -1,0 +1,3 @@
+export { EVENT_GROUP_SELECT, EVENT_RELATION_SELECT, MAJOR_EVENT_SELECT } from './shared';
+export { CsvRow, GraphqlContext, PersonMatch, SubscriptionImportPersonData } from './shared';
+export { EventAttendancesMutationSupport as EventAttendancesResolverBase } from './shared';
