+++ apps/events-backend/src/app/events/attendances/index.ts
@@ -1,0 +1,4 @@
+export { EventAttendanceCsvImportResolver } from './event-attendances.csv-import.resolver';
+export { EventAttendancesMutationsResolver } from './event-attendances.mutations.resolver';
+export { EventAttendancesQueriesResolver } from './event-attendances.queries.resolver';
+export { MajorEventSubscriptionCsvImportResolver } from './major-event-subscription-csv-import.resolver';
