+++ apps/events-backend/src/app/events/attendances/shared/index.ts
@@ -1,0 +1,3 @@
+export { EventAttendancesMutationSupport } from './attendance-mutation-support';
+export { EVENT_GROUP_SELECT, EVENT_RELATION_SELECT, MAJOR_EVENT_SELECT } from './selects';
+export { CsvRow, GraphqlContext, PersonMatch, SubscriptionImportPersonData } from './types';
