+++ apps/events-backend/src/app/events/attendances/shared/selects.ts
@@ -1,0 +1,88 @@
+import { Prisma } from '@prisma/client';
+
+export const MAJOR_EVENT_SELECT = {
+  id: true,
+  name: true,
+  emoji: true,
+  startDate: true,
+  endDate: true,
+  description: true,
+  subscriptionStartDate: true,
+  subscriptionEndDate: true,
+  maxCoursesPerAttendee: true,
+  maxLecturesPerAttendee: true,
+  buttonText: true,
+  buttonLink: true,
+  contactInfo: true,
+  contactType: true,
+  isPaymentRequired: true,
+  shouldIssueCertificateForNonPayingAttendees: true,
+  shouldIssueCertificateForNonSubscribedAttendees: true,
+  additionalPaymentInfo: true,
+  deletedAt: true,
+  createdAt: true,
+  createdById: true,
+  updatedAt: true,
+  updatedById: true,
+} satisfies Prisma.MajorEventSelect;
+
+export const EVENT_GROUP_SELECT = {
+  id: true,
+  name: true,
+  emoji: true,
+  shouldIssueCertificate: true,
+  shouldIssueCertificateForNonPayingAttendees: true,
+  shouldIssueCertificateForNonSubscribedAttendees: true,
+  shouldIssueCertificateForEachEvent: true,
+  shouldIssuePartialCertificate: true,
+  deletedAt: true,
+  createdAt: true,
+  createdById: true,
+  updatedAt: true,
+  updatedById: true,
+} satisfies Prisma.EventGroupSelect;
+
+export const EVENT_RELATION_SELECT = {
+  id: true,
+  name: true,
+  creditMinutes: true,
+  startDate: true,
+  endDate: true,
+  type: true,
+  emoji: true,
+  description: true,
+  shortDescription: true,
+  latitude: true,
+  longitude: true,
+  locationDescription: true,
+  majorEventId: true,
+  majorEvent: {
+    select: MAJOR_EVENT_SELECT,
+  },
+  eventGroupId: true,
+  eventGroup: {
+    select: EVENT_GROUP_SELECT,
+  },
+  allowSubscription: true,
+  subscriptionStartDate: true,
+  subscriptionEndDate: true,
+  slots: true,
+  autoSubscribe: true,
+  shouldIssueCertificate: true,
+  shouldIssueCertificateForNonPayingAttendees: true,
+  shouldIssueCertificateForNonSubscribedAttendees: true,
+  shouldCollectAttendance: true,
+  isOnlineAttendanceAllowed: true,
+  onlineAttendanceCode: true,
+  onlineAttendanceStartDate: true,
+  onlineAttendanceEndDate: true,
+  publiclyVisible: true,
+  youtubeCode: true,
+  buttonText: true,
+  buttonLink: true,
+  deletedAt: true,
+  createdAt: true,
+  createdById: true,
+  updatedAt: true,
+  updatedById: true,
+} satisfies Prisma.EventSelect;
